CORE-5679 - reorder hook assignments so "onConnect" can't be called before onClose is hooked. - improve some logging and error messaging - move ws package
