+++ libs/http-rpc/http-rpc-server-impl/src/main/kotlin/net/corda/httprpc/server/impl/websocket/WsExceptionMapper.kt
@@ -1,4 +1,4 @@
-package net.corda.httprpc.server.impl.apigen.processing.ws
+package net.corda.httprpc.server.impl.websocket
 
 import java.lang.IllegalArgumentException
 import net.corda.httprpc.ws.WebSocketProtocolViolationException
